add typehints to ia and lending

+++ openlibrary/core/lending.py
@@ -1,5 +1,7 @@
 """Module for providing core functionality of lending on Open Library."""
 
+from __future__ import annotations  # Needed for 'Loan' return types early on
+
 import datetime
 import logging
 import time
@@ -24,6 +26,8 @@ from . import ia
 if TYPE_CHECKING:
     from openlibrary.book_providers import EbookAccess
     from openlibrary.plugins.upstream.models import Edition
+
+    from .waitinglist import WaitingLoan
 
 
 logger = logging.getLogger(__name__)
@@ -335,7 +339,7 @@ class AvailabilityStatusV2(AvailabilityStatus):
 
 
 def get_ebook_access_availability(
-    ocaid: str, ebook_access: 'EbookAccess'
+    ocaid: str, ebook_access: EbookAccess
 ) -> AvailabilityStatusV2:
     from openlibrary.book_providers import EbookAccess
 
@@ -565,7 +569,7 @@ def add_availability(
     return items
 
 
-def get_items_and_add_availability(ocaids: list[str]) -> dict[str, "Edition"]:
+def get_items_and_add_availability(ocaids: list[str]) -> dict[str, Edition]:
     """
     Get Editions from OCAIDs and attach their availabiliity.
 
@@ -613,7 +617,7 @@ def is_loaned_out_on_ol(identifier: str) -> bool:
     return bool(loan)
 
 
-def get_loan(identifier, user_key=None):
+def get_loan(identifier: str, user_key: str | None = None):
     """Returns the loan object for given identifier, if a loan exists.
 
     If user_key is specified, it returns the loan only if that user is
@@ -630,12 +634,13 @@ def get_loan(identifier, user_key=None):
     d = web.ctx.site.store.get("loan-" + identifier)
     if d and (
         user_key is None
-        or (d['user'] == account.username)
-        or (d['user'] == account.itemname)
+        or (account and d['user'] == account.username)
+        or (account and d['user'] == account.itemname)
     ):
         loan = Loan(d)
         if loan.is_expired():
-            return loan.delete()
+            loan.delete()
+            return None
     try:
         _loan = _get_ia_loan(identifier, account and userkey2userid(account.username))
     except Exception:  # TODO: Narrow exception scope
@@ -649,12 +654,12 @@ def get_loan(identifier, user_key=None):
     return _loan
 
 
-def _get_ia_loan(identifier: str, userid: str):
+def _get_ia_loan(identifier: str, userid: str | None = None):
     ia_loan = ia_lending_api.get_loan(identifier, userid)
     return ia_loan and Loan.from_ia_loan(ia_loan)
 
 
-def get_loans_of_user(user_key):
+def get_loans_of_user(user_key: str) -> list[Loan]:
     """TODO: Remove inclusion of local data; should only come from IA"""
     if 'env' not in web.ctx:
         """For the get_cached_user_loans to call the API if no cache is present,
@@ -665,7 +670,9 @@ def get_loans_of_user(user_key):
     account = OpenLibraryAccount.get(username=user_key.split('/')[-1])
 
     loandata = web.ctx.site.store.values(type='/type/loan', name='user', value=user_key)
-    loans = [Loan(d) for d in loandata] + (_get_ia_loans_of_user(account.itemname))
+    loans = [Loan(d) for d in loandata]
+    if account and account.itemname:
+        loans += _get_ia_loans_of_user(account.itemname)
     # Set patron's loans in cache w/ now timestamp
     get_cached_loans_of_user.memcache_set(
         [user_key], {}, loans or [], time.time()
@@ -680,7 +687,7 @@ get_cached_loans_of_user = cache.memcache_memoize(
 )
 
 
-def get_user_waiting_loans(user_key):
+def get_user_waiting_loans(user_key: str) -> list[WaitingLoan]:
     """Gets the waitingloans of the patron.
 
     Returns [] if user has no waitingloans.
@@ -692,7 +699,7 @@ def get_user_waiting_loans(user_key):
 
     try:
         account = OpenLibraryAccount.get(key=user_key)
-        itemname = account.itemname
+        itemname = account.itemname if account else None
         result = WaitingLoan.query(userid=itemname)
         get_cached_user_waiting_loans.memcache_set(
             [user_key], {}, result or {}, time.time()
@@ -709,12 +716,14 @@ get_cached_user_waiting_loans = cache.memcache_memoize(
 )
 
 
-def _get_ia_loans_of_user(userid):
+def _get_ia_loans_of_user(userid: str) -> list[Loan]:
     ia_loans = ia_lending_api.find_loans(userid=userid)
     return [Loan.from_ia_loan(d) for d in ia_loans]
 
 
-def create_loan(identifier, resource_type, user_key, book_key=None):
+def create_loan(
+    identifier: str, resource_type: str, user_key: str, book_key: str | None = None
+) -> Loan | None:
     """Creates a loan and returns it."""
     ia_loan = ia_lending_api.create_loan(
         identifier=identifier, format=resource_type, userid=user_key, ol_key=book_key
@@ -725,10 +734,7 @@ def create_loan(identifier, resource_type, user_key, book_key=None):
         eventer.trigger("loan-created", loan)
         sync_loan(identifier)
         return loan
-
-    # loan = Loan.new(identifier, resource_type, user_key, book_key)
-    # loan.save()
-    # return loan
+    return None
 
 
 NOT_INITIALIZED = object()
@@ -798,7 +804,7 @@ def sync_loan(identifier, loan=NOT_INITIALIZED):
 
 class EBookRecord(dict):
     @staticmethod
-    def find(identifier):
+    def find(identifier: str) -> EBookRecord:
         key = "ebooks/" + identifier
         d = web.ctx.site.store.get(key) or {"_key": key, "type": "ebook", "_rev": 1}
         return EBookRecord(d)
@@ -819,7 +825,9 @@ class Loan(dict):
     """Model for loan."""
 
     @staticmethod
-    def new(identifier, resource_type, user_key, book_key=None):
+    def new(
+        identifier: str, resource_type: str, user_key: str, book_key: str | None = None
+    ) -> Loan:
         """Creates a new loan object.
 
         The caller is expected to call save method to save the loan.
@@ -868,7 +876,7 @@ class Loan(dict):
         )
 
     @staticmethod
-    def from_ia_loan(data):
+    def from_ia_loan(data: dict) -> Loan:
         if data['userid'].startswith('ol:'):
             user_key = '/people/' + data['userid'][len('ol:') :]
         elif data['userid'].startswith('@'):
@@ -908,11 +916,11 @@ class Loan(dict):
         }
         return Loan(d)
 
-    def is_ol_loan(self):
+    def is_ol_loan(self) -> bool:
         # self['user'] will be None for IA loans
         return self['user'] is not None
 
-    def save(self):
+    def save(self) -> None:
         # loans stored at IA are not supposed to be saved at OL.
         # This call must have been made in mistake.
         if self.get("stored_at") == "ia":
@@ -923,12 +931,12 @@ class Loan(dict):
         # Inform listers that a loan is created/updated
         eventer.trigger("loan-created", self)
 
-    def is_expired(self):
+    def is_expired(self) -> bool:
         return (
             self['expiry'] and self['expiry'] < datetime.datetime.utcnow().isoformat()
         )
 
-    def is_yet_to_be_fulfilled(self):
+    def is_yet_to_be_fulfilled(self) -> bool:
         """Returns True if the loan is not yet fulfilled and fulfillment time
         is not expired.
         """
@@ -937,7 +945,7 @@ class Loan(dict):
             and (time.time() - self['loaned_at']) < LOAN_FULFILLMENT_TIMEOUT_SECONDS
         )
 
-    def return_loan(self):
+    def return_loan(self) -> bool:
         logger.info("*** return_loan ***")
         if self['resource_type'] == 'bookreader':
             self.delete()
@@ -945,13 +953,13 @@ class Loan(dict):
         else:
             return False
 
-    def delete(self):
+    def delete(self) -> None:
         loan = dict(self, returned_at=time.time())
         user_key = self['user']
         account = OpenLibraryAccount.get(key=user_key)
         if self.get("stored_at") == 'ia':
             ia_lending_api.delete_loan(self['ocaid'], userkey2userid(user_key))
-            if account.itemname:
+            if account and account.itemname:
                 ia_lending_api.delete_loan(self['ocaid'], account.itemname)
         else:
             web.ctx.site.store.delete(self['_key'])
@@ -961,7 +969,7 @@ class Loan(dict):
         eventer.trigger("loan-completed", loan)
 
 
-def resolve_identifier(identifier):
+def resolve_identifier(identifier: str) -> str | None:
     """Returns the OL book key for given IA identifier."""
     if keys := web.ctx.site.things({'type': '/type/edition', 'ocaid': identifier}):
         return keys[0]
@@ -969,12 +977,12 @@ def resolve_identifier(identifier):
         return "/books/ia:" + identifier
 
 
-def userkey2userid(user_key):
+def userkey2userid(user_key: str) -> str:
     username = user_key.split("/")[-1]
     return "ol:" + username
 
 
-def get_resource_id(identifier, resource_type):
+def get_resource_id(identifier: str, resource_type: str) -> str | None:
     """Returns the resource_id for an identifier for the specified resource_type.
 
     The resource_id is found by looking at external_identifiers field in the
@@ -996,6 +1004,7 @@ def get_resource_id(identifier, resource_type):
         acs, rtype, resource_id = eid.split(":", 2)
         if rtype == resource_type:
             return resource_id
+    return None
 
 
 def update_loan_status(identifier):
